requirements.txt
aspy.yaml==1.1.1
atomicwrites==1.2.1
attrs==18.2.0
cached-property==1.5.1
cfgv==1.1.0
click==6.7
Flask==1.0.2
identify==1.1.5
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
nodeenv==1.3.2
pluggy==0.7.1
pre-commit==1.11.0
psycopg2-binary==2.7.5
py==1.6.0
pytest==3.8.0
PyYAML==3.13
six==1.11.0
toml==0.9.6
tox==3.3.0
virtualenv==16.0.0
Werkzeug==0.14.1
